# api/docs/swagger.yaml
basePath: /
definitions:
  EditAccountRequest:
    properties:
      action:
        example: removeAvatar
        type: string
      mimeType:
        example: image/png
        type: string
    required:
    - action
    type: object
  ErrorResponse:
    properties:
      code:
        example: InternalError
        type: string
      error:
        example: An unexpected error occurred
        type: string
    type: object
...
  PresignedUrlResponse:
    properties:
      fields:
        additionalProperties:
          type: string
        type: object
      url:
        type: string
    type: object
...
  User:
    properties:
      accountDeletionSchedule:
        type: string
      avatar:
        example: https://example.com/avatar.png
        type: string
      createdAt:
        example: "2025-07-12T12:11:54.450476-04:00"
        type: string
      displayName:
        example: jane_doe
        type: string
      email:
        example: jane_doe@mail.com
        type: string
      id:
        example: HW4beTVvbTUPRxun9MXZxwKPjmC2
        type: string
      scheduledForDeletionAt:
        type: string
      updatedAt:
        example: "2025-07-15T12:11:54.450476-04:00"
        type: string
    required:
    - displayName
    - email
    - id
    type: object
...
host: localhost:8080
info:
  contact:
    name: Kit
    url: https://github.com/kit-g
  description: A simple fitness tracker API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Heart of Yours API
  version: "1.0"
paths:
  /accounts:
    delete:
      consumes:
      - application/json
      description: Schedules account deletion for the authenticated user
      operationId: deleteAccount
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user account
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Accounts are managed by Firebase so we just need to store them
      operationId: registerAccount
      parameters:
      - description: User request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Creates an account record
      tags:
      - accounts
  /accounts/{accountId}:
    get:
      consumes:
      - application/json
      description: Returns user account information for the authenticated user
      operationId: getAccount
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user account
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: 'Performs various account editing actions: undoAccountDeletion,
        removeAvatar, uploadAvatar'
      operationId: editAccount
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      - description: Edit account request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/EditAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PresignedUrlResponse'
        "204":
          description: No Content
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Edit user account
      tags:
      - accounts
...
swagger: "2.0"